WaveSpectra.jl Documentation

WaveSpectra.jl Documentation

Functions

Index

Functions

WaveSpectra.DiscreteOmnidirectionalSpectrum — Type

DiscreteOmnidirectionalSpectrum(value::AbstractVecOrMat{<:Quantity}, frequency::AbstractVector{<:Quantity}; density::Bool=true)

Create discrete omnidirectional spectrum using two Unitful vectors, or matrix and vector.

Example

julia> using WaveSpectra, Unitful

julia> v1=f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s1 = DiscreteOmnidirectionalSpectrum(v1,f);

julia> v2 = ones(typeof(1u"Hz"), 3, 3)
3×3 Matrix{Quantity{Int64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}}:
 1 Hz  1 Hz  1 Hz
 1 Hz  1 Hz  1 Hz
 1 Hz  1 Hz  1 Hz

julia> s2 = DiscreteOmnidirectionalSpectrum(v2, f);

Unitful.unit — Method

unit(x::Quantity{T,D,U}) where {T,D,U}
unit(x::Type{Quantity{T,D,U}}) where {T,D,U}

Returns the units associated with a Quantity or Quantity type.

Examples:

julia> unit(1.0u"m") == u"m"
true

julia> unit(typeof(1.0u"m")) == u"m"
true

Unitful.unit(::DiscreteOmnidirectionalSpectrum)

Extends from Unitful, returns the units of the spectra vector/matrix in the struct

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> unit(s)
m Hz⁻¹

Unitful.unit(::OmnidirectionalSpectrum{TS, TF})

Extends from Unitful, returns the units of the spectra values.

Example

julia> using WaveSpectra, Unitful

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz"))
(::OmnidirectionalSpectrum{Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}, Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}})     (generic function with 1 method)

julia> unit(s)
Hz

Unitful.dimension — Method

dimension(x::Unit)

Returns a Unitful.Dimensions object describing the given unit x.

dimension(u::Units{U,D}) where {U,D}

Returns a Unitful.Dimensions object corresponding to the dimensions of the units, D. For a dimensionless combination of units, a Unitful.Dimensions{()} object is returned (NoDims).

Examples:

julia> dimension(u"m")
𝐋

julia> typeof(dimension(u"m"))
Unitful.Dimensions{(Unitful.Dimension{:Length}(1//1),)}

julia> dimension(u"m/km")
NoDims

dimension(x::Quantity{T,D}) where {T,D}
dimension(::Type{Quantity{T,D,U}}) where {T,D,U}

Returns a Unitful.Dimensions object D corresponding to the dimensions of quantity x. For a dimensionless Unitful.Quantity, a Unitful.Dimensions{()} object is returned (NoDims).

Examples:

julia> dimension(1.0u"m")
𝐋

julia> typeof(dimension(1.0u"m/μm"))
Unitful.Dimensions{()}

Unitful.dimension(::DiscreteOmnidirectionalSpectrum)

Extends from Unitful, returns the dimension of the spectra vector/matrix in the struct

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> dimension(s)
𝐋 𝐓

Unitful.unit(::OmnidirectionalSpectrum{TS, TF})

Extends from Unitful, returns the dimension of the spectra values.

Example

julia> using WaveSpectra, Unitful

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz"))
(::OmnidirectionalSpectrum{Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}, Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}})     (generic function with 1 method)

julia> dimension(s)
𝐓⁻¹

WaveSpectra.frequency_unit — Function

frequency_unit(::DiscreteOmnidirectionalSpectrum)

Return the units of the frequency vector in the struct

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> frequency_unit(s)
Hz

Unitful.unit(::OmnidirectionalSpectrum{TS, TF})

Extends from Unitful, returns the units of the expected frequency.

Example

julia> using WaveSpectra, Unitful

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz"))
(::OmnidirectionalSpectrum{Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}, Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}})     (generic function with 1 method)

julia> unit(s)
Hz

WaveSpectra.frequency_dimension — Function

frequency_dimension(::DiscreteOmnidirectionalSpectrum)

Return the dimension of the frequency vector in the struct

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> frequency_dimension(s)
𝐓⁻¹

Unitful.unit(::OmnidirectionalSpectrum{TS, TF})

Extends from Unitful, returns the dimension of the expected frequency.

Example

julia> using WaveSpectra, Unitful

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz"))
(::OmnidirectionalSpectrum{Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}, Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}})     (generic function with 1 method)

julia> dimension(s)
𝐓⁻¹

WaveSpectra.quantity — Function

quantity(::DiscreteOmnidirectionalSpectrum{TS, TF, D})

Return the dimensions and units of the product between spectra and frequency.

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> quantity(s)
(𝐋, m)

quantity(::OmnidirectionalSpectrum{TS, TF})

Return the dimensions and units of the product between spectra and frequency.

Example

julia> using WaveSpectra, Unitful

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz"))
(::OmnidirectionalSpectrum{Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}, Quantity{Float64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}})     (generic function with 1 method)

julia> quantity(s)
(𝐓⁻², Hz²)

WaveSpectra.spectral_moment — Function

spectral_moment(spectrum::DiscreteOmnidirectionalSpectrum, n::Real=0; alg::AbstractIntegralAlgorithm=TrapezoidalRule())

Calculate the n-th spectral moment of a discrete spectra struct.

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1u"Hz", 5u"Hz", 5)
(1.0:1.0:5.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> spectral_moment(s, -1)
4.0 s⁻¹

julia> spectral_moment(s, 0)
12.0 s⁻²

julia> spectral_moment(s, 1)
42.0 s⁻³

spectrum::OmnidirectionalSpectrum{TS, TF}, n::Int, f_begin::Union{Quantity, Nothing}=nothing, 
    f_end::Union{Quantity, Nothing}=nothing; alg::AbstractIntegralAlgorithm=QuadGKJL(), kwargs...

Calculate the n-th spectral moment of a discrete spectra struct, use given range if specified.

Example

julia> using WaveSpectra, Unitful

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz"));

julia> spectral_moment(s, -1, 1u"Hz", 5u"Hz")
4.0 s⁻¹

julia> spectral_moment(s, 0, 1u"Hz", 5u"Hz")
12.0 s⁻²

julia> spectral_moment(s, 1, 1u"Hz", 5u"Hz")
41.333333333333336 s⁻³

WaveSpectra.energy_period — Function

energy_period(spectrum::DiscreteOmnidirectionalSpectrum; alg::AbstractIntegralAlgorithm=TrapezoidalRule())

Calculate the energy period of a discrete spectra struct.

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1u"Hz", 5u"Hz", 5)
(1.0:1.0:5.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> energy_period(s)
0.3333333333333333 s

WaveSpectra.significant_waveheight — Function

significant_waveheight(spectrum::DiscreteOmnidirectionalSpectrum; alg::AbstractIntegralAlgorithm=TrapezoidalRule())

Calculate the significant waveheight of a discrete spectra struct.

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1u"Hz", 5u"Hz", 5)
(1.0:1.0:5.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> significant_waveheight(s)
13.856406460551018 s⁻¹

WaveSpectra.convert_frequency — Function

convert_frequency(spectrum::DiscreteOmnidirectionalSpectrum{TS, TF}, TF_new, dispersion::Dispersion=Dispersion()) where {TS, TF}

Converts the spectra into the new frequency units using the DimensionfulAngles.Dispersion relation and returns a new struct with updated spectrum and frequency.

See also DimensionfulAngles.Dispersion

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1u"Hz", 5u"Hz", 5)
(1.0:1.0:5.0) Hz

julia> s1 = DiscreteOmnidirectionalSpectrum(v,f);

julia> s2 = convert_frequency(s1, typeof(1u"Hz^-1"));

Index

WaveSpectra.DiscreteOmnidirectionalSpectrum

Unitful.dimension

Unitful.unit

WaveSpectra.convert_frequency

WaveSpectra.energy_period

WaveSpectra.frequency_dimension

WaveSpectra.frequency_unit

WaveSpectra.quantity

WaveSpectra.significant_waveheight

WaveSpectra.spectral_moment