WaveSpectra.jl Documentation

WaveSpectra.jl Documentation

Functions

Index

Functions

WaveSpectra.DiscreteOmnidirectionalSpectrum — Type

DiscreteOmnidirectionalSpectrum(value::AbstractVecOrMat{<:Quantity}, frequency::AbstractVector{<:Quantity}; density::Bool=true)

Create discrete omnidirectional spectrum using two Unitful vectors, or matrix and vector.

Example

julia> using WaveSpectra, Unitful

julia> v1=f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s1 = DiscreteOmnidirectionalSpectrum(v1,f);

julia> v2 = ones(typeof(1u"Hz"), 3, 3)
3×3 Matrix{Quantity{Int64, 𝐓⁻¹, Unitful.FreeUnits{(Hz,), 𝐓⁻¹, nothing}}}:
 1 Hz  1 Hz  1 Hz
 1 Hz  1 Hz  1 Hz
 1 Hz  1 Hz  1 Hz

julia> s2 = DiscreteOmnidirectionalSpectrum(v2, f);

WaveSpectra.OmnidirectionalSpectrum — Type

OmnidirectionalSpectrum(func::Function, TS::DataType, TF::DataType)

Create an omnidirectional spectrum using a a function with the appropriate data types.

Example

julia> using WaveSpectra, Unitful

julia> s1 = OmnidirectionalSpectrum(x -> x, typeof(1.0u"Hz"), typeof(1.0u"Hz"));

WaveSpectra.frequency_unit — Function

frequency_unit(::DiscreteOmnidirectionalSpectrum)

Return the units of the frequency vector.

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> frequency_unit(s)
Hz

frequency_unit(::OmnidirectionalSpectrum{TS, TF})

Return the units of the expected frequency vector.

Example

julia> using WaveSpectra, Unitful

julia> s = OmnidirectionalSpectrum(x -> x, typeof(1.0u"Hz"));

julia> unit(s)
Hz

WaveSpectra.frequency_dimension — Function

frequency_dimension(::DiscreteOmnidirectionalSpectrum)

Return the dimension of the frequency vector.

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> frequency_dimension(s)
𝐓⁻¹

frequency_dimension(::OmnidirectionalSpectrum{TS, TF})

Return the dimension of the expected frequency vector.

Example

julia> using WaveSpectra, Unitful

julia> s = OmnidirectionalSpectrum(x -> x, typeof(1.0u"Hz"));

julia> dimension(s)
𝐓⁻¹

WaveSpectra.quantity — Function

quantity(::DiscreteOmnidirectionalSpectrum{TS, TF, D})

Return the dimensions and units of the product between spectra and frequency.

Example

julia> using WaveSpectra, Unitful

julia> v=range(1u"m/Hz", 3u"m/Hz", 3)
(1.0:1.0:3.0) m Hz⁻¹

julia> f=range(1u"Hz", 3u"Hz", 3)
(1.0:1.0:3.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> quantity(s)
(𝐋, m)

quantity(::OmnidirectionalSpectrum{TS, TF})

Return the dimensions and units of the product between spectra and frequency.

Example

julia> using WaveSpectra, Unitful

julia> s = OmnidirectionalSpectrum(x -> x, typeof(1.0u"Hz"));

julia> quantity(s)
(𝐓⁻², Hz²)

WaveSpectra.spectral_moment — Function

spectral_moment(spectrum::DiscreteOmnidirectionalSpectrum, n::Real=0; alg::AbstractIntegralAlgorithm=TrapezoidalRule())

Calculate the n-th spectral moment of a discrete spectra.

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1u"Hz", 5u"Hz", 5)
(1.0:1.0:5.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> spectral_moment(s, -1)
4.0 s⁻¹

julia> spectral_moment(s, 0)
12.0 s⁻²

julia> spectral_moment(s, 1)
42.0 s⁻³

spectral_moment(spectrum::OmnidirectionalSpectrum{TS, TF}, n::Int,
f_begin::Union{Quantity, Nothing}=nothing, f_end::Union{Quantity, Nothing}=nothing;
alg::AbstractIntegralAlgorithm=QuadGKJL(), kwargs...)

Calculate the n-th spectral moment of a spectra, use range if given.

Example

julia> using WaveSpectra, Unitful

julia> s = OmnidirectionalSpectrum(x -> x, typeof(1.0u"Hz"));

julia> spectral_moment(s, -1, 1u"Hz", 5u"Hz")
4.0 s⁻¹

julia> spectral_moment(s, 0, 1u"Hz", 5u"Hz")
12.0 s⁻²

julia> spectral_moment(s, 1, 1u"Hz", 5u"Hz")
41.333333333333336 s⁻³

WaveSpectra.energy_period — Function

energy_period(
    spectrum::OmnidirectionalSpectrum{TS,TF},
    f_begin::Union{Quantity,Nothing}=nothing, f_end::Union{Quantity,Nothing}=nothing;
    alg::AbstractIntegralAlgorithm=QuadGKJL(), kwargs...
) where {TS,TF<:Frequency}

Calculate the energy period of an omnidirectional spectra between a range if given.

Example

julia> using WaveSpectra, Unitful

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz"));

julia> energy_period(s, 1u"Hz", 5u"Hz")
0.3333333333333333 s

energy_period(spectrum::OmnidirectionalSpectrum{TS,TF},
    dispersion::Dispersion=Dispersion(), f_begin::Union{Quantity,Nothing}=nothing,
    f_end::Union{Quantity,Nothing}=nothing; alg::AbstractIntegralAlgorithm=QuadGKJL(),
    kwargs...
) where {TS,TF}

Calculate the energy period of an omnidirectional spectra in Hz.

Example

julia> using WaveSpectra, DimensionfulAngles

julia> using Unitful:@u_str

julia> func(x) = x
func (generic function with 1 method)

julia> s = OmnidirectionalSpectrum(func, typeof(1.0u"Hz^-1"));

julia> energy_period(s, Dispersion(), 1u"Hz", 5u"Hz")
0.688888888888889 s

energy_period(spectrum::DiscreteOmnidirectionalSpectrum; alg::AbstractIntegralAlgorithm=TrapezoidalRule())

Calculate the energy period of a discrete spectra.

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1u"Hz", 5u"Hz", 5)
(1.0:1.0:5.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> energy_period(s)
0.3333333333333333 s

WaveSpectra.significant_waveheight — Function

significant_waveheight(spectrum::OmnidirectionalSpectrum{TS,TF},
    f_begin::Union{Quantity,Nothing}=nothing, f_end::Union{Quantity,Nothing}=nothing;
    alg::AbstractIntegralAlgorithm=QuadGKJL(), kwargs...
) where {TS,TF}

Calculate the significant waveheight of an omnidirectional spectra.

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1,5,5)
1.0:1.0:5.0

julia> v = v.*u"m^2/Hz"
(1.0:1.0:5.0) m² Hz⁻¹

julia> f = f.*u"Hz"
(1.0:1.0:5.0) Hz

julia> s = OmnidirectionalSpectrum(v,f);

julia> significant_waveheight(s)
13.856406451197893 m

significant_waveheight(spectrum::DiscreteOmnidirectionalSpectrum; alg::AbstractIntegralAlgorithm=TrapezoidalRule())

Calculate the significant waveheight of a discrete spectra.

Example

julia> using WaveSpectra, Unitful

julia> v=f=range(1u"Hz", 5u"Hz", 5)
(1.0:1.0:5.0) Hz

julia> s = DiscreteOmnidirectionalSpectrum(v,f);

julia> significant_waveheight(s)
13.856406460551018 s⁻¹

WaveSpectra.convert_frequency — Function

convert_frequency(spectrum::OmnidirectionalSpectrum{TS,TF}, TF_new, dispersion::Dispersion=Dispersion())

Converts the spectra into the new frequency units using the DimensionfulAngles.Dispersion relation and returns a new struct with an updated function.

See also DimensionfulAngles.Dispersion

Example

julia> using WaveSpectra, Unitful

julia> s1 = OmnidirectionalSpectrum(x -> x, typeof(1.0u"Hz"));

julia> s2 = convert_frequency(s1, 1.0u"Hz^-1");

Index

WaveSpectra.DiscreteOmnidirectionalSpectrum

WaveSpectra.OmnidirectionalSpectrum

WaveSpectra.convert_frequency

WaveSpectra.energy_period

WaveSpectra.frequency_dimension

WaveSpectra.frequency_unit

WaveSpectra.quantity

WaveSpectra.significant_waveheight

WaveSpectra.spectral_moment